src/VerdeVida.Weather/requirements.txt
flask==3.0.3
gunicorn==21.2.0

Flask==3.0.3
opentelemetry-distro
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-flask
gunicorn


